started with simple electron beam option

+++ Energy Distribution Model/src/MCMC.h
@@ -19,11 +19,11 @@
 
 struct MCMC_Parameters
 {
-	int numberSamples = 3e5;
+	int numberSamples = (int)3e5;
 	int burnIn = 1000;
 	int lag = 30;
-	float proposalSigma[3] = { 0.005, 0.005, 0.2 };
-	int seed = std::time(0);
+	float proposalSigma[3] = { 0.005f, 0.005f, 0.2f };
+	int seed = (int)std::time(0);
 
 	std::string String();
 };
